Initial commit
A reasonable draft of a Visual Studio name demangling implementation.
There are about 25% more lines of code than in the Wine implementation,
but this version is more accurate (for our test suite), much easier to
understand, and is more flexible.  For example, this code could easily
be expanded to mangle as well. Many bugs remain to be fixed and many
improvements could still be made.