Convert DemangledType to JSON, infrastructure.
These routines convert from DemangledType to JSON.  The main conversion
routine, JsonOutput::convert(), is not yet finished.  JsonOutput::raw(),
however, is complete and working, and provides a complete output of the
contents of the DemangledType data structure.  The raw output was
created as a debugging and exploration tool in the process of figuring
out how to create a more user-friendly output in convert().