Add JSON output for the demangle program.
This adds JSON output for demangle.  It is only as good as demangle_json
is.  Updates to the JSON output will likely be in the demangle_json
code, not in this code.

There is a hidden --raw option that outputs all the data from the
DemangledType in the same form as the DemangledType object.
( demangle --json --raw )