Make {Array,Object}::add easier to use.
Using the original API, in order to add, say, and int to an array, you
had to use array.add(builder.simple(4)).  One would like to be able
to say array.add(4) instead, but we don't want to add a proliferation of
virtual methods for every simple data type.

So the Simple object was created.  This object has constructors for each
simple data type.  An Array or Object implementation needs to override a
method that takes a Simple refref as an argument, and it can use the
Simple object along with its own builder to convert to a NodeRef.  The
users data type (say, int) will be implicitly converted to a Simple
object using Simple's conversion constructors, and everything will just
work.

This makes using the library much simpler, at the cost of it being more
complicated to understand how the library works.