Flatten the name data structure.
This is a breaking change, but is the right thing to do.  Now entries in
the name tend not to be layered down an extra level.  This makes the
data structure much simpler and easier to use, but I now have to fix all
the places in the code that depend on the old, messy approach.